+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/ProjectTemplate/CS4HA/Startup.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Threading.Tasks;
 using hhnl.HomeAssistantNet.Automations.Automation;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace CS4HA
 {
@@ -9,6 +11,11 @@ namespace CS4HA
         public static Task Main(string[] args)
         {
             return new Startup().RunAsync(args);
+        }
+
+        protected override void ConfigureServices(HostBuilderContext builderContext, IServiceCollection serviceCollection)
+        {
+            // You can register custom services here.
         }
     }
 }