+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.Automations/Automation/Runner/AutomationRunnerFactory.cs
@@ -24,6 +24,7 @@ namespace hhnl.HomeAssistantNet.Automations.Automation.Runner
                 ReentryPolicy.Allow => new AllowAutomationRunner(entry, _serviceProvider),
                 ReentryPolicy.Discard => new DiscardAutomationRunner(entry, _serviceProvider),
                 ReentryPolicy.Queue => new QueueAutomationRunner(entry, _serviceProvider),
+                ReentryPolicy.QueueLatest => new QueueLatestAutomationRunner(entry, _serviceProvider),
                 ReentryPolicy.CancelPrevious => new CancelPreviousAutomationRunner(entry, _serviceProvider),
                 _ => throw new ArgumentOutOfRangeException(nameof(entry))
             };