+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.Automations/Automation/StateChangedNotificationHandler.cs
@@ -27,12 +27,12 @@ namespace hhnl.HomeAssistantNet.Automations.Automation
             // Update entities
             await _entityRegistry.UpdateEntityAsync(notification.EntityId, notification.NewState);
 
-            // Run automations
             var automations = _automationRegistry.GetAutomationsTrackingEntity(notification.EntityId);
 
+            // Run automations
             foreach (var automation in automations)
             {
-                await _automationService.EnqueueAutomationForEntityChangeAsync(automation, notification.EntityId);
+                await _automationService.EnqueueAutomationForEntityChangeAsync(automation, notification.EntityId, notification.SourceEvent);
             }
         }
     }