+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.Shared/Automation/SnapshotAttribute.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace hhnl.HomeAssistantNet.Shared.Automation
+{
+    /// <summary>
+    /// When applied to a parameter, the entity will always be a snapshot created at the start of the automation.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Parameter)]
+    public class SnapshotAttribute : Attribute
+    {
+    }
+}