+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.Shared/SourceGenerator/IEntitySnapshotProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace hhnl.HomeAssistantNet.Shared.SourceGenerator
+{
+    public interface IEntitySnapshotProvider
+    {
+        T GetSnapshot<T>();
+    }
+}